gov-info-backend/requirements.txt
uvicorn~=0.32.0
fastapi~=0.115.4
psycopg2~=2.9.10
pydantic~=2.9.2
pandas~=2.2.3
python-dotenv~=1.0.1
langchain-openai~=0.2.6
pydantic-settings~=2.6.1
SQLAlchemy~=2.0.36
tiktoken~=0.8.0
langchain-community~=0.3.11
openai~=1.59.8
requests~=2.32.3